refactor: Create index file in animation module for easier import.
