+++ src/animation/index.js
@@ -1,0 +1,5 @@
+/* @flow */
+export { default as AnimatedComponent } from './AnimatedComponent';
+export { default as AnimatedRotateComponent } from './AnimatedRotateComponent';
+export { default as AnimatedScaleComponent } from './AnimatedScaleComponent';
+export { default as AnimatedSlideComponent } from './AnimatedSlideComponent';
